Better organisation and some refactoring and some comments
