+++ functionary/train/packing/assets/correct_packing_attention.png
